icon_creation_tool/requirements.txt
google-generativeai
Pillow
python-dotenv
requests
